[refactor] Split oversized files to satisfy file_length (#182)
Splits the four design-refresh VCs (Settings, AlarmFiring, CreateAlarm,
Onboarding) and AudioService into sibling extension files so SwiftLint's
file_length / type_body_length / function_body_length warnings stay clear.

Pure refactor — no behaviour changes. Observers, animation timing, side
effects, and threading are all preserved verbatim.

* SettingsViewController: extracted +Sections, +TransactionCell, +Legal;
  cellForRowAt split into focused per-section helpers.
* AlarmFiringViewController: extracted +Layout (view-stack + constraint
  builders) and +ViewLifecycle (clock, glow, snooze handler, top-up sheet).
* CreateAlarmViewController: extracted +Sections (per-row factories +
  cell registration) and +Pickers (sound/volume/theme push handlers,
  delete confirmation).
* OnboardingViewController: extracted +Pages (page builders, preset stack,
  CTA actions, pager-state). setupUI split into chrome / target wiring /
  page-stack helpers.
* AudioService: extracted +Tone (synthetic tone generator + WAV packer +
  binary helpers); startAlarmSound split into resolveAlarmPlayer,
  configurePlayerVolume, handleStartWhileNonStopped helpers.
* Views/DesignSystem/SectionHeaderView: new shared caps-styled header used
  by CreateAlarm; SettingsVC keeps its private header per #182's
  'do NOT touch' note.

Fixes #182
