+++ SnoozePay/SnoozePay/ViewControllers/Settings/SettingsViewController.swift
@@ -19,8 +19,10 @@ class SettingsViewController: UIViewController {
     }()
 
     /// Held weakly so the cell may be recycled (and the label deallocated)
-    /// without leaving the observer with a dangling reference.
-    private weak var balanceAmountLabel: UILabel?
+    /// without leaving the observer with a dangling reference. `internal`
+    /// so the cross-file `+Sections.makeBalanceRow` can capture a fresh
+    /// reference each time the row is dequeued (#182).
+    weak var balanceAmountLabel: UILabel?
 
     // MARK: - Init
 
@@ -180,7 +182,7 @@ extension SettingsViewController: UITableViewDataSource {
         guard let title = self.tableView(tableView, titleForHeaderInSection: section) else {
             return nil
         }
-        return SettingsSectionHeaderView(text: title)
+        return makeSettingsSectionHeader(text: title)
     }
 
     func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
@@ -196,178 +198,42 @@ extension SettingsViewController: UITableViewDataSource {
         guard let section = Section(rawValue: indexPath.section) else { return UITableViewCell() }
 
         switch section {
-        case .account:
-            if indexPath.row == 0 {
-                return makeIconRow(
-                    systemName: "list.bullet.rectangle",
-                    iconColor: AppColors.info500,
-                    title: "История транзакций",
-                    accessory: .disclosureIndicator
-                )
-            } else {
-                let cell = makeIconRow(
-                    systemName: "dollarsign.circle",
-                    iconColor: AppColors.money500,
-                    title: "Баланс",
-                    accessory: .none
-                )
-
-                let balanceAmount = UILabel()
-                balanceAmount.text = "₽\(Int(BalanceService.shared.balance))"
-                balanceAmount.font = AppTypography.moneyMd
-                balanceAmount.textColor = AppColors.money500
-                balanceAmount.translatesAutoresizingMaskIntoConstraints = false
-                balanceAmount.setContentHuggingPriority(.required, for: .horizontal)
-                cell.contentView.addSubview(balanceAmount)
-
-                NSLayoutConstraint.activate([
-                    balanceAmount.trailingAnchor.constraint(
-                        equalTo: cell.contentView.trailingAnchor,
-                        constant: -AppSpacing.lg
-                    ),
-                    balanceAmount.centerYAnchor.constraint(equalTo: cell.contentView.centerYAnchor)
-                ])
-
-                // Track the latest label instance so the balance observer can
-                // refresh it. When the cell is recycled the weak ref nils out
-                // automatically, so the observer becomes a no-op until the row
-                // is rebuilt — at which point this assignment captures the new label.
-                self.balanceAmountLabel = balanceAmount
-
-                cell.selectionStyle = .none
-                return cell
-            }
-
-        case .referral:
-            return makeReferralCell(at: indexPath)
-
-        case .appearance:
-            // Dedicated cell type owns the segment control — avoids the previous
-            // removeFromSuperview/re-add dance that left the control bound to
-            // whichever cell rendered last.
-            guard let cell = tableView.dequeueReusableCell(
-                withIdentifier: ThemeSegmentCell.reuseID,
-                for: indexPath
-            ) as? ThemeSegmentCell else {
-                assertionFailure("dequeueReusableCell returned wrong type for \(ThemeSegmentCell.reuseID)")
-                return UITableViewCell()
-            }
-            cell.configure(selectedIndex: themeSegmentIndex) { [weak self] index in
-                self?.handleThemeSegmentChange(index)
-            }
-            return cell
-
-        case .info:
-            let title = indexPath.row == 0 ? "Политика конфиденциальности" : "Пользовательское соглашение"
-            return makeIconRow(
-                systemName: "doc.text",
-                iconColor: UIColor.systemGray,
-                title: title,
-                accessory: .disclosureIndicator
-            )
-
-        case .contact:
-            let cell = UITableViewCell(style: .default, reuseIdentifier: nil)
-            cell.backgroundColor = .secondarySystemBackground
-
-            let icon = makeSettingsIcon(systemName: "envelope", backgroundColor: AppColors.warn500)
-            let titleLabel = UILabel()
-            titleLabel.text = "Связаться с нами"
-            titleLabel.font = AppTypography.bodyLg
-            titleLabel.textColor = AppColors.fg1
-
-            let detailLabel = UILabel()
-            detailLabel.text = "support@alarmcash.app"
-            detailLabel.font = AppTypography.meta
-            detailLabel.textColor = AppColors.fg3
-
-            let textStack = UIStackView(arrangedSubviews: [titleLabel, detailLabel])
-            textStack.axis = .vertical
-            textStack.spacing = 2
-
-            let stack = UIStackView(arrangedSubviews: [icon, textStack])
-            stack.axis = .horizontal
-            stack.spacing = AppSpacing.md
-            stack.alignment = .center
-            stack.translatesAutoresizingMaskIntoConstraints = false
-            cell.contentView.addSubview(stack)
-
-            NSLayoutConstraint.activate([
-                stack.leadingAnchor.constraint(
-                    equalTo: cell.contentView.leadingAnchor,
-                    constant: AppSpacing.lg
-                ),
-                stack.centerYAnchor.constraint(equalTo: cell.contentView.centerYAnchor),
-                stack.trailingAnchor.constraint(
-                    lessThanOrEqualTo: cell.contentView.trailingAnchor,
-                    constant: -AppSpacing.lg
-                )
-            ])
-
-            cell.accessoryType = .disclosureIndicator
-            return cell
+        case .account:    return makeAccountCell(at: indexPath)
+        case .referral:   return makeReferralCell(at: indexPath)
+        case .appearance: return makeAppearanceCell(tableView, at: indexPath)
+        case .info:       return makeInfoRow(at: indexPath)
+        case .contact:    return makeContactRow()
         }
     }
 
-    // MARK: - Cell factory helpers
-
-    private func makeIconRow(
-        systemName: String,
-        iconColor: UIColor,
-        title: String,
-        accessory: UITableViewCell.AccessoryType
-    ) -> UITableViewCell {
-        let cell = UITableViewCell(style: .default, reuseIdentifier: nil)
-        cell.backgroundColor = .secondarySystemBackground
-
-        let icon = makeSettingsIcon(systemName: systemName, backgroundColor: iconColor)
-        let label = UILabel()
-        label.text = title
-        label.font = AppTypography.bodyLg
-        label.textColor = AppColors.fg1
-
-        let stack = UIStackView(arrangedSubviews: [icon, label])
-        stack.axis = .horizontal
-        stack.spacing = AppSpacing.md
-        stack.alignment = .center
-        stack.translatesAutoresizingMaskIntoConstraints = false
-        cell.contentView.addSubview(stack)
-
-        NSLayoutConstraint.activate([
-            stack.leadingAnchor.constraint(equalTo: cell.contentView.leadingAnchor, constant: AppSpacing.lg),
-            stack.centerYAnchor.constraint(equalTo: cell.contentView.centerYAnchor)
-        ])
-
-        cell.accessoryType = accessory
-        return cell
+    /// Account section dispatches between the History row (`row == 0`) and
+    /// the Balance row (`row == 1`). Split off `cellForRowAt` so the main
+    /// switch stays a one-screen overview (#182).
+    private func makeAccountCell(at indexPath: IndexPath) -> UITableViewCell {
+        indexPath.row == 0 ? makeTransactionHistoryRow() : makeBalanceRow()
     }
 
-    /// Creates a rounded-rect icon with an SF Symbol, similar to iOS Settings style
-    private func makeSettingsIcon(systemName: String, backgroundColor: UIColor) -> UIView {
-        let size: CGFloat = 30
-        let container = UIView()
-        container.backgroundColor = backgroundColor
-        container.layer.cornerRadius = 7
-        container.layer.masksToBounds = true
-        container.translatesAutoresizingMaskIntoConstraints = false
-
-        let imageView = UIImageView()
-        imageView.image = UIImage(systemName: systemName)?.withConfiguration(
-            UIImage.SymbolConfiguration(pointSize: 15, weight: .medium)
-        )
-        imageView.tintColor = .white
-        imageView.contentMode = .scaleAspectFit
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        container.addSubview(imageView)
-
-        NSLayoutConstraint.activate([
-            container.widthAnchor.constraint(equalToConstant: size),
-            container.heightAnchor.constraint(equalToConstant: size),
-            imageView.centerXAnchor.constraint(equalTo: container.centerXAnchor),
-            imageView.centerYAnchor.constraint(equalTo: container.centerYAnchor)
-        ])
-
-        return container
+    /// Dequeue + configure for the segmented Theme cell. Pulled out of
+    /// `cellForRowAt` so the main switch stays under SwiftLint's
+    /// `function_body_length` cap (#182).
+    private func makeAppearanceCell(
+        _ tableView: UITableView,
+        at indexPath: IndexPath
+    ) -> UITableViewCell {
+        // Dedicated cell type owns the segment control — avoids the previous
+        // removeFromSuperview/re-add dance that left the control bound to
+        // whichever cell rendered last.
+        guard let cell = tableView.dequeueReusableCell(
+            withIdentifier: ThemeSegmentCell.reuseID,
+            for: indexPath
+        ) as? ThemeSegmentCell else {
+            assertionFailure("dequeueReusableCell returned wrong type for \(ThemeSegmentCell.reuseID)")
+            return UITableViewCell()
+        }
+        cell.configure(selectedIndex: themeSegmentIndex) { [weak self] index in
+            self?.handleThemeSegmentChange(index)
+        }
+        return cell
     }
 }
 
@@ -453,364 +319,8 @@ extension SettingsViewController: UITableViewDelegate {
     }
 }
 
-// MARK: - Transaction History VC
-
-final class TransactionHistoryViewController: UIViewController {
-
-    private let tableView: UITableView = {
-        let table = UITableView(frame: .zero, style: .insetGrouped)
-        table.translatesAutoresizingMaskIntoConstraints = false
-        return table
-    }()
-
-    private var transactions: [Transaction] = []
-    private let alarmRepository = AlarmRepository.shared
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        title = "История транзакций"
-        view.backgroundColor = AppColors.bg0
-        setupTableView()
-        loadTransactions()
-    }
-
-    private func setupTableView() {
-        view.addSubview(tableView)
-        tableView.dataSource = self
-        tableView.delegate = self
-        tableView.register(TransactionCell.self, forCellReuseIdentifier: TransactionCell.reuseID)
-        NSLayoutConstraint.activate([
-            tableView.topAnchor.constraint(equalTo: view.topAnchor),
-            tableView.leadingAnchor.constraint(equalTo: view.leadingAnchor),
-            tableView.trailingAnchor.constraint(equalTo: view.trailingAnchor),
-            tableView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
-        ])
-    }
-
-    private func loadTransactions() {
-        // Use the checked variant so a corrupt ledger surfaces an alert
-        // instead of rendering the misleading "Нет транзакций" empty state
-        // — that would let the user assume the app reset itself (issue #117).
-        do {
-            transactions = try TransactionRepository.shared.fetchAllChecked()
-        } catch let error as TransactionRepository.RepositoryError {
-            transactions = []
-            presentLoadError(error)
-        } catch {
-            transactions = []
-        }
-        tableView.reloadData()
-    }
-
-    private func presentLoadError(_ error: LocalizedError) {
-        let alert = UIAlertController(
-            title: "Ошибка",
-            message: error.errorDescription,
-            preferredStyle: .alert
-        )
-        alert.addAction(UIAlertAction(title: "OK", style: .default))
-        present(alert, animated: true)
-    }
-}
-
-extension TransactionHistoryViewController: UITableViewDataSource, UITableViewDelegate {
-
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        transactions.isEmpty ? 1 : transactions.count
-    }
-
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        if transactions.isEmpty {
-            let cell = UITableViewCell(style: .default, reuseIdentifier: nil)
-            cell.textLabel?.text = "Нет транзакций"
-            cell.textLabel?.textColor = .secondaryLabel
-            cell.textLabel?.textAlignment = .center
-            cell.backgroundColor = .secondarySystemBackground
-            cell.selectionStyle = .none
-            return cell
-        }
-
-        guard let cell = tableView.dequeueReusableCell(
-            withIdentifier: TransactionCell.reuseID, for: indexPath
-        ) as? TransactionCell else {
-            assertionFailure("dequeueReusableCell returned wrong type for \(TransactionCell.reuseID)")
-            return UITableViewCell()
-        }
-
-        let transaction = transactions[indexPath.row]
-
-        // Look up alarm name if available — best-effort enrichment. If the
-        // alarm store is corrupt we log and fall back to nil so the row
-        // still renders (transaction amount/date is still meaningful).
-        // Without this, a single corrupt alarms blob would silently strip
-        // alarm names from every history row (issue #117).
-        var alarmName: String?
-        if let alarmIDString = transaction.alarmID, let alarmUUID = UUID(uuidString: alarmIDString) {
-            do {
-                alarmName = try alarmRepository.fetchChecked(id: alarmUUID)?.name
-            } catch {
-                let errorDesc = String(describing: error)
-                AppLogger.ui.error(
-                    "TxHistory: name lookup failed for \(alarmUUID, privacy: .private): \(errorDesc, privacy: .public)"
-                )
-            }
-        }
-
-        cell.configure(with: transaction, alarmName: alarmName)
-        return cell
-    }
-
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        transactions.isEmpty ? 52 : 64
-    }
-
-    func tableView(_ tableView: UITableView, titleForHeaderInSection section: Int) -> String? {
-        "ИСТОРИЯ"
-    }
-
-    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
-        guard let title = self.tableView(tableView, titleForHeaderInSection: section) else {
-            return nil
-        }
-        return SettingsSectionHeaderView(text: title)
-    }
-
-    func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
-        UITableView.automaticDimension
-    }
-
-    /// Mirror SettingsViewController so the transaction list reads as a card
-    /// in light mode rather than blending into the page.
-    func tableView(_ tableView: UITableView, willDisplay cell: UITableViewCell, forRowAt indexPath: IndexPath) {
-        let totalRows = self.tableView(tableView, numberOfRowsInSection: indexPath.section)
-        let position = CardRowPosition.resolve(row: indexPath.row, totalRows: totalRows)
-        cell.styleAsCardRow(position: position)
-    }
-}
-
-// MARK: - Transaction Cell
-
-final class TransactionCell: UITableViewCell {
-
-    static let reuseID = "TransactionCell"
-
-    // MARK: - UI
-
-    private let iconContainer: UIView = {
-        let view = UIView()
-        view.layer.cornerRadius = 20
-        view.layer.masksToBounds = true
-        view.translatesAutoresizingMaskIntoConstraints = false
-        return view
-    }()
-
-    private let iconImageView: UIImageView = {
-        let imageView = UIImageView()
-        imageView.tintColor = .white
-        imageView.contentMode = .scaleAspectFit
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        return imageView
-    }()
-
-    private let titleLabel: UILabel = {
-        let label = UILabel()
-        label.font = AppTypography.bodyLg
-        label.textColor = AppColors.fg1
-        return label
-    }()
-
-    private let subtitleLabel: UILabel = {
-        let label = UILabel()
-        label.font = AppTypography.meta
-        label.textColor = AppColors.fg3
-        return label
-    }()
-
-    private let amountLabel: UILabel = {
-        let label = UILabel()
-        label.font = AppTypography.moneyMd
-        label.textAlignment = .right
-        label.setContentHuggingPriority(.required, for: .horizontal)
-        label.setContentCompressionResistancePriority(.required, for: .horizontal)
-        return label
-    }()
-
-    // MARK: - Init
-
-    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
-        super.init(style: style, reuseIdentifier: reuseIdentifier)
-        setupUI()
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    // MARK: - Setup
-
-    private func setupUI() {
-        backgroundColor = .secondarySystemBackground
-        selectionStyle = .none
-
-        iconContainer.addSubview(iconImageView)
-
-        let textStack = UIStackView(arrangedSubviews: [titleLabel, subtitleLabel])
-        textStack.axis = .vertical
-        textStack.spacing = 2
-
-        let mainStack = UIStackView(arrangedSubviews: [iconContainer, textStack, amountLabel])
-        mainStack.axis = .horizontal
-        mainStack.spacing = AppSpacing.md
-        mainStack.alignment = .center
-        mainStack.translatesAutoresizingMaskIntoConstraints = false
-        contentView.addSubview(mainStack)
-
-        NSLayoutConstraint.activate([
-            iconContainer.widthAnchor.constraint(equalToConstant: 40),
-            iconContainer.heightAnchor.constraint(equalToConstant: 40),
-            iconImageView.centerXAnchor.constraint(equalTo: iconContainer.centerXAnchor),
-            iconImageView.centerYAnchor.constraint(equalTo: iconContainer.centerYAnchor),
-            iconImageView.widthAnchor.constraint(equalToConstant: 18),
-            iconImageView.heightAnchor.constraint(equalToConstant: 18),
-
-            mainStack.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: AppSpacing.lg),
-            mainStack.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -AppSpacing.lg),
-            mainStack.topAnchor.constraint(equalTo: contentView.topAnchor, constant: AppSpacing.sm),
-            mainStack.bottomAnchor.constraint(equalTo: contentView.bottomAnchor, constant: -AppSpacing.sm)
-        ])
-    }
-
-    // MARK: - Configure
-
-    func configure(with transaction: Transaction, alarmName: String?) {
-        let isCredit = transaction.type != .charge
-
-        // Icon — credits (topup, promotion) read green/up; charges red/down.
-        // Promotion gets a distinct gift glyph so the user can tell a
-        // referral bonus apart from an IAP top-up at a glance (issue #144).
-        iconContainer.backgroundColor = isCredit ? AppColors.money500 : AppColors.pain500
-        let iconName: String
-        switch transaction.type {
-        case .topup:     iconName = "arrow.up"
-        case .charge:    iconName = "arrow.down"
-        case .promotion: iconName = "gift.fill"
-        }
-        iconImageView.image = UIImage(systemName: iconName)?.withConfiguration(
-            UIImage.SymbolConfiguration(pointSize: 14, weight: .bold)
-        )
-
-        // Title — promotion reads "Бонус" so the row doesn't claim the user
-        // paid for the credit (which would mislead anyone scanning their
-        // history for IAP receipts).
-        switch transaction.type {
-        case .topup:     titleLabel.text = "Пополнение"
-        case .charge:    titleLabel.text = "Откладывание"
-        case .promotion: titleLabel.text = "Бонус"
-        }
-
-        // Subtitle: alarm name + date for charges; just the date otherwise.
-        let dateString = Self.formatDate(transaction.createdAt)
-        if let name = alarmName, transaction.type == .charge {
-            subtitleLabel.text = "\(name) \u{00B7} \(dateString)"
-        } else {
-            subtitleLabel.text = dateString
-        }
-
-        // Amount
-        let amount = Int(transaction.amount)
-        if isCredit {
-            amountLabel.text = "+₽\(amount)"
-            amountLabel.textColor = AppColors.money500
-        } else {
-            amountLabel.text = "₽\(amount)"
-            amountLabel.textColor = AppColors.pain500
-        }
-    }
-
-    // MARK: - Date formatting
-
-    private static let dateFormatter: DateFormatter = {
-        let formatter = DateFormatter()
-        formatter.locale = Locale(identifier: "ru_RU")
-        formatter.dateFormat = "d MMMM 'в' HH:mm"
-        return formatter
-    }()
-
-    private static func formatDate(_ date: Date) -> String {
-        dateFormatter.string(from: date)
-    }
-}
-
-// MARK: - Section header
-
-/// Caps-styled section header used by both `SettingsViewController` and the
-/// nested `TransactionHistoryViewController`. Mirrors `CreateAlarmViewController`'s
-/// private `SectionHeaderView` so the brand `caps` role + 0.12em tracking lands
-/// consistently across the Settings stack (#177).
-private final class SettingsSectionHeaderView: UIView {
-
-    init(text: String) {
-        super.init(frame: .zero)
-        backgroundColor = .clear
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.attributedText = NSAttributedString(
-            string: text.uppercased(),
-            attributes: [
-                .font: AppTypography.caps,
-                .kern: AppTypography.capsKerning,
-                .foregroundColor: AppColors.fg3
-            ]
-        )
-        addSubview(label)
-        NSLayoutConstraint.activate([
-            label.leadingAnchor.constraint(equalTo: layoutMarginsGuide.leadingAnchor),
-            label.trailingAnchor.constraint(lessThanOrEqualTo: layoutMarginsGuide.trailingAnchor),
-            label.topAnchor.constraint(equalTo: topAnchor, constant: AppSpacing.lg),
-            label.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -AppSpacing.sm)
-        ])
-    }
-
-    @available(*, unavailable)
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-}
-
-// MARK: - Legal VC (simple text placeholder)
-
-final class LegalViewController: UIViewController {
-
-    private let legalTitle: String
-
-    init(title: String) {
-        self.legalTitle = title
-        super.init(nibName: nil, bundle: nil)
-    }
-
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        title = legalTitle
-        view.backgroundColor = AppColors.bg0
-
-        let textView = UITextView()
-        textView.text = "\(legalTitle)\n\nДокумент будет добавлен перед публикацией в App Store."
-        textView.font = AppTypography.body
-        textView.textColor = AppColors.fg1
-        textView.backgroundColor = .clear
-        textView.isEditable = false
-        textView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(textView)
-
-        NSLayoutConstraint.activate([
-            textView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 16),
-            textView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
-            textView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16),
-            textView.bottomAnchor.constraint(equalTo: view.bottomAnchor)
-        ])
-    }
-}
+// `TransactionHistoryViewController`, `TransactionCell`,
+// `SettingsSectionHeaderView`, and `LegalViewController` were extracted to
+// sibling extension files (`+TransactionCell.swift`, `+Sections.swift`,
+// `+Legal.swift`) in #182 to satisfy SwiftLint's `file_length` cap. Behaviour
+// is unchanged — only the physical location moved.
