+++ SnoozePay/SnoozePay/Views/DesignSystem/SectionHeaderView.swift
@@ -1,0 +1,36 @@
+import UIKit
+
+/// Shared caps-styled section header used by inset-grouped tables across the
+/// design-refresh surfaces (CreateAlarm, Settings, Transaction history). Mirrors
+/// the original private `SectionHeaderView` introduced in #143 — extracted to
+/// `Views/DesignSystem` so the brand `caps` role + 0.12em tracking lands
+/// consistently without duplicate per-screen copies (#182).
+final class SectionHeaderView: UIView {
+
+    init(text: String) {
+        super.init(frame: .zero)
+        backgroundColor = .clear
+        let label = UILabel()
+        label.translatesAutoresizingMaskIntoConstraints = false
+        label.attributedText = NSAttributedString(
+            string: text.uppercased(),
+            attributes: [
+                .font: AppTypography.caps,
+                .kern: AppTypography.capsKerning,
+                .foregroundColor: AppColors.fg3
+            ]
+        )
+        addSubview(label)
+        NSLayoutConstraint.activate([
+            label.leadingAnchor.constraint(equalTo: layoutMarginsGuide.leadingAnchor),
+            label.trailingAnchor.constraint(lessThanOrEqualTo: layoutMarginsGuide.trailingAnchor),
+            label.topAnchor.constraint(equalTo: topAnchor, constant: AppSpacing.lg),
+            label.bottomAnchor.constraint(equalTo: bottomAnchor, constant: -AppSpacing.sm)
+        ])
+    }
+
+    @available(*, unavailable)
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
